refactor(chrome-ext): move Google Docs logic to its own module
